Add missing closing tag for span
Modern browsers have an error correction system.
Therefore, when the HTML code is not completely correct,
they deal with the correct interpretation.
However, the Markdown engine does not have such a mechanism.
If the HTML is written incorrectly,
Markdown will try to process it in the own way.

In the previous commit I changed the HTML file to Markdown.
But after starting the blog locally,
I noticed that the tags page looked different.
Incorrect tags have been escaped.

It turned out that the closing tag was missing.
Without it, Markdown treated other closing tags as quotes
and escaped them.

Additionally, Markdown can be sensitive to indentation.
In this case, the indentation caused the HTML code to be interpreted
as preformatted source code.

Here is an example:

```
<div class="language-plaintext highlighter-rouge"><div class="highlight"><pre class="highlight"><code>&lt;h2 id="Rails Girls" style="color: #c91b26"&gt;Rails Girls&lt;/h2&gt;
&lt;ul class="post-list"&gt;
		&lt;li&gt;&lt;a href="/rails-girls-bielsko-biala"&gt;Rails Girls Bielsko-Biała (7 March 2020)&lt;span class="entry-date"&gt;&lt;time datetime="2020-05-21T00:00:00-04:00" itemprop="datePublished"&gt;May 21, 2020&lt;/time&gt;&lt;/span&gt;&lt;/a&gt;&lt;/li&gt;

```

Fixing the indentation and adding a closing tag for span,
everything looks correct.