Add missing closing tag for span in categories
In the previous commit I changed the HTML file to Markdown.
But after starting the blog locally,
I noticed that the tags page looked different.
Incorrect tags have been escaped.

It turned out that the closing tag was missing.
Without it, Markdown treated other closing tags as quotes
and escaped them.

Here is an example:

```
</time>&lt;/a&gt;&lt;/li&gt;
```

After making the correction, everything works as it should.